+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Added more merchant defined fields
 - Added Payer Auth Services
+- Send neighborhood as district
 
 ## [1.3.2] - 2022-07-11